refactor/args: make args part of config module
